php: Simplify the end-user installation with a helper script
Besides making the instructions simpler, this also improves the robustness of
the Docker image, because we can adjust filenames and directory structure as
necessary.

The Debian Dockerfile has been synchronized with the Alpine one, embedded the
SHA-256 hash of the regular downloads, instead of going through `apt`, as this
simplifies the placement in the top-level `/tideways/` directory without
unecessary duplication or complicated cleanup.

The updated instructions are now as follows:

Debian:

    COPY --from=ghcr.io/tideways/php:latest /tideways/ /tideways/
    RUN docker-php-ext-enable --ini-name tideways.ini $(php /tideways/get-ext-path.php)

Alpine:

    COPY --from=ghcr.io/tideways/php:alpine /tideways/ /tideways/
    RUN docker-php-ext-enable --ini-name tideways.ini $(php /tideways/get-ext-path.php)

The previous instructions are no longer valid, but they have not yet been
documented.