Rebuild the images daily
Caches in GitHub Actions live for at most 1 week, a weekly rebuild thus might
not reliably see cache hits.

Rebuild the image daily, with caching enabled, it should actually result in a
new image and it will pick up rebuilds of the base image (in case of the daemon
image) more quickly.